Fix bug when weight is set to 0
There was a slight chance that a feature could be enabled even when
weight was set to 0.

Not sure how to test this reliably. If I put a test in to test the edge
case, it's possible that future updates to the code will silently break
it again. Hopefully this comment is enough.